// backend-curriculum/src/common/interfaces/login.ts
export interface AuthenticatedUser {
  uuid: string;
  username: string;
  email: string;
}

export interface AuthenticatedUserResponse {
  message?: string;
  user: AuthenticatedUser;
}

export interface UserData extends AuthenticatedUser {
  password?: string;
  isActive?: boolean;
}